remove license
I wanted to earn money somehow,
but I don't have a concrete plan how...

Figuring it out with people might be helpful.

Some plan was to have a license that is more like
Pay for this if you appreciate the work put in.
9euros/per year | 18 euros/biyearly.

But it would not be required to have a lincese in order to use it.
Maybe dev versions should require a lincese... And have a stable one that is free...

Do I think that this story will happen... I am highly skeptical,
but I do see value in this.